src/requirements.txt
openai==1.3.8 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.2 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.24.0.post1
requests
fastapi==0.105.0
python-dotenv